Clean up handles and fields
