+++ src/Microsoft.Diagnostics.Runtime/src/Implementation/ClrmdPrimitiveType.cs
@@ -31,7 +31,7 @@ namespace Microsoft.Diagnostics.Runtime.Implementation
         public override ClrModule Module { get; }
         public override ClrElementType ElementType { get; }
         public override bool IsShared => false;
-        public override int StaticSize => ClrmdField.GetSize(this, ElementType);
+        public override int StaticSize => ClrField.GetSize(this, ElementType);
         public override ClrType? BaseType => null; // todo;
         public override ClrHeap Heap { get; }
         public override IEnumerable<ClrInterface> EnumerateInterfaces() => Enumerable.Empty<ClrInterface>();
